+++ packages/swirl-components/src/components.d.ts
@@ -177,6 +177,10 @@ export namespace Components {
           * Expands an accordion item.
          */
         "expandItem": (itemId: string) => Promise<void>;
+        /**
+          * @default false
+         */
+        "multiExpand": boolean;
     }
     interface SwirlAccordionItem {
         /**
@@ -1851,6 +1855,13 @@ export namespace Components {
         "size": SwirlIconSize;
     }
     interface SwirlIconFormatListBulleted {
+        "color"?: SwirlIconColor1;
+        /**
+          * @default 24
+         */
+        "size": SwirlIconSize;
+    }
+    interface SwirlIconForward {
         "color"?: SwirlIconColor1;
         /**
           * @default 24
@@ -6459,6 +6470,12 @@ declare global {
         prototype: HTMLSwirlIconFormatListBulletedElement;
         new (): HTMLSwirlIconFormatListBulletedElement;
     };
+    interface HTMLSwirlIconForwardElement extends Components.SwirlIconForward, HTMLStencilElement {
+    }
+    var HTMLSwirlIconForwardElement: {
+        prototype: HTMLSwirlIconForwardElement;
+        new (): HTMLSwirlIconForwardElement;
+    };
     interface HTMLSwirlIconFullscreenElement extends Components.SwirlIconFullscreen, HTMLStencilElement {
     }
     var HTMLSwirlIconFullscreenElement: {
@@ -8973,6 +8990,7 @@ declare global {
         "swirl-icon-folder-shared": HTMLSwirlIconFolderSharedElement;
         "swirl-icon-format-h-one": HTMLSwirlIconFormatHOneElement;
         "swirl-icon-format-list-bulleted": HTMLSwirlIconFormatListBulletedElement;
+        "swirl-icon-forward": HTMLSwirlIconForwardElement;
         "swirl-icon-fullscreen": HTMLSwirlIconFullscreenElement;
         "swirl-icon-fullscreen-exit": HTMLSwirlIconFullscreenExitElement;
         "swirl-icon-gif": HTMLSwirlIconGifElement;
@@ -9308,6 +9326,10 @@ declare namespace LocalJSX {
     interface FileManager {
     }
     interface SwirlAccordion {
+        /**
+          * @default false
+         */
+        "multiExpand"?: boolean;
         "onExpandedItemChange"?: (event: SwirlAccordionCustomEvent<string>) => void;
     }
     interface SwirlAccordionItem {
@@ -10920,6 +10942,13 @@ declare namespace LocalJSX {
         "size"?: SwirlIconSize;
     }
     interface SwirlIconFormatListBulleted {
+        "color"?: SwirlIconColor1;
+        /**
+          * @default 24
+         */
+        "size"?: SwirlIconSize;
+    }
+    interface SwirlIconForward {
         "color"?: SwirlIconColor1;
         /**
           * @default 24
@@ -14176,6 +14205,7 @@ declare namespace LocalJSX {
         "swirl-icon-folder-shared": SwirlIconFolderShared;
         "swirl-icon-format-h-one": SwirlIconFormatHOne;
         "swirl-icon-format-list-bulleted": SwirlIconFormatListBulleted;
+        "swirl-icon-forward": SwirlIconForward;
         "swirl-icon-fullscreen": SwirlIconFullscreen;
         "swirl-icon-fullscreen-exit": SwirlIconFullscreenExit;
         "swirl-icon-gif": SwirlIconGif;
@@ -14656,6 +14686,7 @@ declare module "@stencil/core" {
             "swirl-icon-folder-shared": LocalJSX.SwirlIconFolderShared & JSXBase.HTMLAttributes<HTMLSwirlIconFolderSharedElement>;
             "swirl-icon-format-h-one": LocalJSX.SwirlIconFormatHOne & JSXBase.HTMLAttributes<HTMLSwirlIconFormatHOneElement>;
             "swirl-icon-format-list-bulleted": LocalJSX.SwirlIconFormatListBulleted & JSXBase.HTMLAttributes<HTMLSwirlIconFormatListBulletedElement>;
+            "swirl-icon-forward": LocalJSX.SwirlIconForward & JSXBase.HTMLAttributes<HTMLSwirlIconForwardElement>;
             "swirl-icon-fullscreen": LocalJSX.SwirlIconFullscreen & JSXBase.HTMLAttributes<HTMLSwirlIconFullscreenElement>;
             "swirl-icon-fullscreen-exit": LocalJSX.SwirlIconFullscreenExit & JSXBase.HTMLAttributes<HTMLSwirlIconFullscreenExitElement>;
             "swirl-icon-gif": LocalJSX.SwirlIconGif & JSXBase.HTMLAttributes<HTMLSwirlIconGifElement>;