+++ packages/swirl-tokens/dart/lib/styles.light.dart
@@ -4,7 +4,7 @@
 //
 
 // Do not edit directly
-// Generated on Wed, 03 Dec 2025 12:35:52 GMT
+// Generated on Thu, 04 Dec 2025 14:20:20 GMT